change location of Test
