+++ TestFlashLFQ/Test.cs
@@ -111,5 +111,49 @@ namespace Test
             Assert.That(int1 == int3);
             Assert.That(int1 == int5);
         }
+        [Test]
+        public static void MergeFlashLFQResults()
+        {
+            SpectraFileInfo rawA = new SpectraFileInfo(Path.Combine(TestContext.CurrentContext.TestDirectory, @"sliced-raw.raw"), "a", 0, 0, 0);
+            SpectraFileInfo mzmlA = new SpectraFileInfo(Path.Combine(TestContext.CurrentContext.TestDirectory, @"sliced-mzml.mzml"), "a", 0, 1, 0);
+
+            // create some PSMs
+            var pgA = new ProteinGroup("MyProtein", "gene", "org");
+            Identification id1A = new Identification(rawA, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.12193, 2, new List<ProteinGroup> { pgA });
+            Identification id2A = new Identification(rawA, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.05811, 2, new List<ProteinGroup> { pgA });
+            Identification id3A = new Identification(mzmlA, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.12193, 2, new List<ProteinGroup> { pgA });
+            Identification id4A = new Identification(mzmlA, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.05811, 2, new List<ProteinGroup> { pgA });
+
+            // create the FlashLFQ engine
+            FlashLFQEngine engineA = new FlashLFQEngine(new List<Identification> { id1A, id2A, id3A, id4A });
+
+            // run the engine
+            var resultsA = engineA.Run();
+
+            var tempResultsA = resultsA;
+
+            SpectraFileInfo rawB = new SpectraFileInfo(Path.Combine(TestContext.CurrentContext.TestDirectory, @"sliced-rawb.raw"), "b", 0, 0, 0);
+            SpectraFileInfo mzmlB = new SpectraFileInfo(Path.Combine(TestContext.CurrentContext.TestDirectory, @"sliced-mzmlb.mzml"), "b", 0, 1, 0);
+
+            // create some PSMs
+            var pgB = new ProteinGroup("MyProtein", "gene", "org");
+            Identification id1 = new Identification(rawB, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.12193, 2, new List<ProteinGroup> { pgB });
+            Identification id2 = new Identification(rawB, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.05811, 2, new List<ProteinGroup> { pgB });
+            Identification id3 = new Identification(mzmlB, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.12193, 2, new List<ProteinGroup> { pgB });
+            Identification id4 = new Identification(mzmlB, "EGFQVADGPLYR", "EGFQVADGPLYR", 1350.65681, 94.05811, 2, new List<ProteinGroup> { pgB });
+
+            // create the FlashLFQ engine
+            FlashLFQEngine engineB = new FlashLFQEngine(new List<Identification> { id1, id2, id3, id4 });
+
+            // run the engine
+            var resultsB = engineB.Run();
+
+            resultsA.MergeResultsWith(resultsB);
+            Assert.AreEqual(4, resultsA.peaks.Count);
+            Assert.AreEqual(1, resultsA.peptideBaseSequences.Count);
+            Assert.AreEqual(1, resultsA.peptideModifiedSequences.Count);
+            Assert.AreEqual(1, resultsA.proteinGroups.Count);
+            Assert.AreEqual(4, resultsA.spectraFiles.Count);
+        }
     }
 }
